requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

sqlalchemy[asyncio]==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
asyncpg==0.31.0

pydantic[email]==2.9.2
pydantic-settings==2.13.0
python-dotenv==1.0.1
python-multipart==0.0.9

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

httpx==0.27.2

celery[redis]==5.4.0
redis==5.1.1

python-json-logger==2.0.7
slowapi==0.1.9